# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da GateWay API GymApp",
    "description": "Documentação da gateway api de suporte à aplicação GymApp",
    "contact": { "name": "Teste", "email": "teste@teste.com" },
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:2900", "description": "API de teste" }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Login do user",
        "description": "Esta rota permite realizar o login e autenticação do user",
        "tags": ["Autenticação"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              },
              "examples": {
                "Aluno": {
                  "value": {
                    "email": "biancasilva@gmail.com",
                    "password": "passwd"
                  }
                },
                "Admin": {
                  "value": { "email": "admin@admin.com", "password": "admin" }
                },
                "Treinador": {
                  "value": {
                    "email": "treinador@treinador.com",
                    "password": "treinador"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "token": { "type": "string" },
                    "refreshToken": { "type": "string" }
                  }
                },
                "examples": {
                  "Adicionar Musculo": {
                    "value": {
                      "message": "Login concluido com sucesso",
                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2NDk4NTMzOTgsImV4cCI6MTY0OTg1NDI5OCwic3ViIjoiMDEwOGFiNWItZGM1Mi00MGFjLWIwNzctOTFjNjhmZTQ2MjljIn0.rOcVCmv9tDmXI2e59j5Ki0Bv-osf2xH3EukFhnalDtE",
                      "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2NDk4NTMzOTgsImV4cCI6MTY4MTQxMDk5OCwic3ViIjoiMDEwOGFiNWItZGM1Mi00MGFjLWIwNzctOTFjNjhmZTQ2MjljIn0.fMrt4XJoopSHoVeciN8tre0nS62NTLB2iavKF9WKXQM"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/token": {
      "post": {
        "summary": "Gerar novo token de sessão",
        "description": "Esta rota permite gerar um nove token de sessão para um utilizador",
        "tags": ["Autenticação"],
        "security": [{ "basicAuth": [] }],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "token": { "type": "string" } }
                },
                "examples": {
                  "Adicionar Musculo": {
                    "value": {
                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2NDk4NTgxNDQsImV4cCI6MTY0OTg1OTA0NCwic3ViIjoiMDEwOGFiNWItZGM1Mi00MGFjLWIwNzctOTFjNjhmZTQ2MjljIn0.a-zcCM6kOOmTy2WbzwpoTHZ2iamuwb-PXr8N9X6r4cc"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "delete": {
        "summary": "Realizar o logout",
        "description": "Esta rota permite gerar um novo token de sessão para um utilizador",
        "tags": ["Autenticação"],
        "security": [{ "basicAuth": [] }],
        "responses": {
          "500": { "description": "ERROR" },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "msg": { "type": "string" } }
                },
                "examples": {
                  "Adicionar Musculo": {
                    "value": { "msg": "Logout com sucesso" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "basicAuth": { "type": "http", "scheme": "bearer" } },
...
  }
}